MAINT: Move compute_metrics into metrics
