+++ src/spectral_recovery/__init__.py
@@ -6,5 +6,5 @@ processing indices, and computing recovery metrics.
 
 from spectral_recovery.io.raster import read_timeseries
 from spectral_recovery.io.polygon import read_restoration_polygons, read_reference_polygons
-from spectral_recovery.restoration import compute_metrics
+from spectral_recovery.metrics import compute_metrics
 from spectral_recovery.plotting import plot_spectral_trajectory
